Extract common webpack config
